requirements.txt
Pillow==9.5.0
pywin32;platform_system=="Windows"
win32printing;platform_system=="Windows"
pynput~=1.7.6
requests~=2.31.0
keyboard~=0.13.5
pyinstaller
schedule